Tag most of the the Browser Upload tests as AWSv2 specific.
Altough most of the Browser Upload tests are truly specific
to AWSv2, they were being executed also in AWSv4 runs causing
many false-positives. The patch tag them appropriately and
prevents from being executed when S3_USE_SIGV4=1 is specified.
